Invert logic for connecting stdout, stdin, stderr
If I understand the intention correctly, the `stdout`, `stdin`, `stderr` boolean options are meant to be `True` in order to connect these streams of the subprocess to the parent process. This behavior was not working, and this change is meant to fix that.

The Python [subprocess](https://docs.python.org/3/library/subprocess.html) docs explain that using the `None` value will achieve connecting the streams. The existing code only used `None` when the passed in option was `True` (and `os.devnull` otherwise). Therefore, inverting the logic should result in the intended behavior.